Refactor all but league
