+++ lib/season_statistics_module.rb
@@ -1,109 +1,4 @@
-require 'pry'
 module SeasonStatistics
-
-  
-
-  def season_gather #helper
-    seasons =[]
-      @games.each do |game|
-        seasons << game.season
-      end
-    seasons.uniq
-  end
-
-  def hit_helper(season)#helper
-    hash = Hash.new{|h,k| h[k] = [] }
-     season_gather.each do |season|
-      @games.each do |game|
-        if game.season == season
-          hash[season].push(game.game_id)
-        end
-      end
-    end
-    team_hits = Hash.new(0)
-      hash[season].each do |game_i|
-        @game_teams.each do |game|
-        if game_i == game.game_id
-          team_hits[game.team_id] += game.hits
-        end
-      end
-    end
-  team_hits
-  end
-
-  def name_finder(something) #helper
-    @teams.each do |team|
-      if team.team_id == something[1]
-        return team.team_name
-      end
-    end
-  end
- 
-  def season_helper(season, season_type)
-    season = @game_teams.select{|game| season[0..3] == game.game_id[0..3]}
-    season_hash = Hash.new{|team, results| team[results] = {wins: 0, total: 0}}
-    season.each do |game|
-      if game.game_id[5] == season_type
-        season_hash[game.team_id][:wins] += 1 if game.won == 'true'
-        season_hash[game.team_id][:total] += 1
-      end
-    end
-    season_hash
-  end
-
-  def coach_info_helper(season)
-    season = @game_teams.select{|game| season[0..3] == game.game_id[0..3]}
-    coach_info = Hash.new{|coach, results| coach[results] = {wins: 0, total: 0}}
-    season.each do |game|
-      coach_info[game.head_coach][:wins] += 1 if game.won == 'true'
-      coach_info[game.head_coach][:total] += 1
-    end
-    coach_info
-  end
-
-  def team_info_helper(season)
-    season = @game_teams.select{|game| season[0..3] == game.game_id[0..3]}
-    team_info = Hash.new do |results, team|
-      results[team] = {
-        goals: 0,
-        shots: 0,
-        hits:  0,
-        pim:   0,
-        powerPlayOpportunities: 0,
-        powerPlayGoals: 0,
-        giveaways: 0,
-        takeaways: 0}
-    end
-    season.each do |game|
-      team_info[game.team_id][:goals] += game.goals
-      team_info[game.team_id][:shots] += game.shots
-      team_info[game.team_id][:hits] += game.hits
-      team_info[game.team_id][:pim] += game.pim
-      team_info[game.team_id][:powerPlayOpportunities] += game.power_play_opportunities
-      team_info[game.team_id][:powerPlayGoals] += game.power_play_goals
-      team_info[game.team_id][:giveaways] += game.giveaways
-      team_info[game.team_id][:takeaways] += game.takeaways
-    end
-    team_info
-  end
-
-  def season_type_difference(season)
-    reg = season_helper(season, "2")
-    playoff = season_helper(season, "3")
-    array = [reg, playoff]
-    keys = playoff.keys
-    difference = Hash[keys.zip(array.map do |reg_hash|
-      reg_hash.values_at(*keys)
-    end.inject do |a, b|
-      a.zip(b).map do |x, y|
-        if x[:total] != 0 && y[:total] != 0
-          (((x[:wins] + y[:wins]) / (x[:total] + y[:total].to_f)) - (x[:wins] / x[:total].to_f))
-        end
-      end
-    end)]
-    difference
-  end
-
   def biggest_bust(season)
     difference = season_type_difference(season)
     id = difference.compact.min_by {|k,v| v}
@@ -136,17 +31,96 @@ module SeasonStatistics
 
   def most_hits(season)
     highest = hit_helper(season).invert.max
-    name_finder(highest)
+    return_team_name(highest[1])
   end
 
   def fewest_hits(season)
     fewest = hit_helper(season).invert.min
-    name_finder(fewest)
+    return_team_name(fewest[1])
   end
 
   def power_play_goal_percentage(season)
     power_play_goals = team_info_helper(season).sum {|k,v| v[:powerPlayGoals]}.to_f
     total_goals = team_info_helper(season).sum {|k,v| v[:goals]}.to_f
     (power_play_goals / total_goals).round(2)
+  end
+
+  def season_gather
+    @games.map {|game| game.season}.uniq
+  end
+
+  def hit_helper(input_season)
+    hash = Hash.new{|h,k| h[k] = [] }
+     season_gather.each do |season|
+      @games.each do |game|
+        hash[season].push(game.game_id) if game.season == season
+      end
+    end
+    team_hits = Hash.new(0)
+      hash[input_season].each do |game_i|
+        @game_teams.each do |game|
+        team_hits[game.team_id] += game.hits if game_i == game.game_id
+      end
+    end
+  team_hits
+  end
+
+  def season_helper(season, season_type)
+    season = @game_teams.select{|game| season[0..3] == game.game_id[0..3]}
+    season_hash = Hash.new{|team, results| team[results] = {wins: 0, total: 0}}
+    season.each do |game|
+      if game.game_id[5] == season_type
+        season_hash[game.team_id][:wins] += 1 if game.won == 'true'
+        season_hash[game.team_id][:total] += 1
+      end
+    end
+    season_hash
+  end
+
+  def coach_info_helper(season)
+    season = @game_teams.select{|game| season[0..3] == game.game_id[0..3]}
+    coach_info = Hash.new{|coach, results| coach[results] = {wins: 0, total: 0}}
+    season.each do |game|
+      coach_info[game.head_coach][:wins] += 1 if game.won == 'true'
+      coach_info[game.head_coach][:total] += 1
+    end
+    coach_info
+  end
+
+  def team_info_helper(season)
+    season = @game_teams.select{|game| season[0..3] == game.game_id[0..3]}
+    team_info = Hash.new do |results, team|
+      results[team] = {
+        goals: 0,
+        shots: 0,
+        powerPlayGoals: 0}
+    end
+    season.each do |game|
+      team_info[game.team_id][:goals] += game.goals
+      team_info[game.team_id][:shots] += game.shots
+      team_info[game.team_id][:powerPlayGoals] += game.power_play_goals
+    end
+    team_info
+  end
+
+  def season_type_difference(season)
+    reg = season_helper(season, "2")
+    playoff = season_helper(season, "3")
+    array = [reg, playoff]
+    keys = playoff.keys
+    difference = Hash[keys.zip(array.map do |reg_hash|
+      reg_hash.values_at(*keys)
+    end.inject do |reg_keys, playoff_keys|
+      reg_keys.zip(playoff_keys).map do |reg_values, playoff_keys|
+        if reg_values[:total] != 0 && playoff_keys[:total] != 0
+          all_record = reg_values[:wins] + playoff_keys[:wins]
+          all_games  = reg_values[:total] + playoff_keys[:total].to_f
+          regular_season_record = reg_values[:wins] / reg_values[:total].to_f
+
+          all_record / all_games - regular_season_record
+        end
+      end
+    end)]
+    difference
   end
 end
