+++ test/season_statistics_test.rb
@@ -3,7 +3,6 @@ SimpleCov.start
 require 'minitest/autorun'
 require 'minitest/pride'
 require './lib/stat_tracker'
-require 'pry'
 
 class SeasonStatisticsTest < Minitest::Test
   def setup
@@ -42,34 +41,12 @@ class SeasonStatisticsTest < Minitest::Test
     assert_equal "Penguins", @stat_tracker.least_accurate_team("20122013")
   end
 
-
-
   def test_most_hits
-    game_path = './test/data/dummy_game.csv'
-    team_path = './test/data/team_info.csv'
-    game_teams_path = './data/game_teams_stats.csv'
-
-    locations = {games: game_path,
-                 teams: team_path,
-                 game_teams: game_teams_path}
-
-    stat_tracker = StatTracker.from_csv(locations)
-    assert_equal "Oilers", stat_tracker.fewest_hits("20132014")
-    assert_equal "Ducks", stat_tracker.most_hits("20142015")
+    assert_equal "Bruins", @stat_tracker.most_hits("20122013")
   end
 
   def test_fewest_hits
-    game_path = './test/data/dummy_game.csv'
-    team_path = './test/data/team_info.csv'
-    game_teams_path = './data/game_teams_stats.csv'
-
-    locations = {games: game_path,
-                 teams: team_path,
-                 game_teams: game_teams_path}
-
-    stat_tracker = StatTracker.from_csv(locations)
-    assert_equal "Oilers", stat_tracker.fewest_hits("20132014")
-    assert_equal "Hurricanes", stat_tracker.fewest_hits("20142015")
+    assert_equal "Blackhawks", @stat_tracker.fewest_hits("20122013")
   end
 
   def test_power_play_goal_percentage
